+++ corpus/special_characters.clj
@@ -1,0 +1,29 @@
+; This Source Code Form is subject to the terms of the Mozilla Public
+; License, v. 2.0. If a copy of the MPL was not distributed with this
+; file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+@a
+
+@(a)
+
+#(+ 1 %)
+
+#=(+ 1 2)
+
+#"a"
+
+#'a
+
+#'(a b)
+
+`a
+
+`(a b)
+
+~a
+
+~(a b)
+
+~@a
+
+~@(a b)